Add show all regions to dotplot view
